RAG/self-rag/pyproject.toml
[project]
name = "self-rag"
version = "0.1.0"
description = "Self-RAG implementation using Google ADK (Agent Development Kit)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3==1.37.37",
    "faiss-cpu==1.10.0",
    "google-adk==0.5.0",
    "langchain==0.3.23",
    "langchain-aws==0.2.21",
    "langchain-chroma>=0.1.2",
    "langchain-community==0.3.21",
    "litellm>=1.67.0",
    "llama-index==0.12.31",
    "pypdf==5.5.0",
    "python-dotenv==1.1.0",
]
